mcp/requirements.txt
mcp[cli]
splitwise
python-dotenv
python-dateutil
starlette
uvicorn